them du thu

+++ src/main/java/Control/Ad_AddProductControl.java
@@ -37,22 +37,30 @@ public class Ad_AddProductControl extends HttpServlet {
         String maSP;
         SanPham sanPham;
         int maDM;
+        String title;
 
         if (action.equals("delete")) {
             maSP = request.getParameter("maSP");
             SanPhamDAO sanPhamDAO = new SanPhamDAO();
             sanPhamDAO.deleteSP(maSP);
+
+            String mess ="Xóa thành công";
+            HttpSession session = request.getSession();
+            session.setAttribute("getAlert",mess);
+
             response.sendRedirect("Ad_ProductControl");
         } else {
             if (action.equals("modify")) {
                 maSP = request.getParameter("maSP");
                 SanPhamDAO sanPhamDAO = new SanPhamDAO();
                 sanPham = sanPhamDAO.getProductById(Integer.parseInt(maSP));
+                title = "Sửa sản phẩm";
             } else {
                 sanPham = new SanPham();
                 if (maLoai == null || maLoai.equals("")) {
                     maLoai = "1";
                 }
+                title = "Thêm sản phẩm mới";
             }
             LoaispDAO loaispDAO = new LoaispDAO();
             List<LoaiSP> listLoaiSP = loaispDAO.getAllloaisp();
@@ -60,6 +68,7 @@ public class Ad_AddProductControl extends HttpServlet {
 //            List<DanhMuc> listDanhMuc = danhMucDAO.getDanhMucByMaLoai(maLoai);
             List<DanhMuc> listDanhMuc = danhMucDAO.getAllDanhMuc();
 
+            request.setAttribute("title",title);
             request.setAttribute("sanPham", sanPham);
             request.setAttribute("maLoai", maLoai);
             request.setAttribute("listDanhMuc", listDanhMuc);
@@ -86,19 +95,19 @@ public class Ad_AddProductControl extends HttpServlet {
         String oldImage = request.getParameter("oldImage");
         String motangan = request.getParameter("motangan");
         String anh;
+        String mess;
 
         // get images
         Collection<Part> fileParts = request.getParts();
         int index = 0;
         for (int i = index; i < fileParts.size(); i++) {
             Part part = (Part) fileParts.toArray()[i];
-            if (!part.getName().equals("multiPartServlet")) {
+            if (!part.getName().equals("multiPartServlet") || part.getSubmittedFileName() == null || part.getSubmittedFileName().equals("")) {
                 System.out.println(part.getName());
                 fileParts.remove(part);
                 i = index - 1;
             }
         }
-
         // create a folder containing images
         LoaispDAO loaispDAO = new LoaispDAO();
         String tenLoai = loaispDAO.getLoaispByMaDM(maDM).trim().replace("/", "");
@@ -114,8 +123,8 @@ public class Ad_AddProductControl extends HttpServlet {
         String formattedDate = myDateObj.format(myFormatObj);
 
         String path = "uploads" + "/" + tenLoai + "/" + tenDM + "/" + "date " + formattedDate;
-
         List<String> images = new ArrayList<>();
+
         if (!fileParts.isEmpty()) {
 
             String realPath = request.getServletContext().getRealPath("/" + path);
@@ -133,33 +142,64 @@ public class Ad_AddProductControl extends HttpServlet {
             anh = images.get(0);
         } else {
             anh = oldImage;
+            if(oldImage.equals(""))
+            {
+                anh = "asset/img/images/emptyImage.png";
+            }
         }
         SanPhamDAO sanPhamDAO = new SanPhamDAO();
         if (maSP.equals("") || maSP == null || maSP.equals("0")) {
             sanPhamDAO.addSanPham(maDM, tensanpham, motasanpham, giagoc, giabanthuong, giakhuyenmai, soluong, anh, motangan);
             SanPham sanPham = sanPhamDAO.getNewSP();
             AnhSPDAO anhSPDAO = new AnhSPDAO();
-            for (String image : images) {
-                anhSPDAO.addAnhSP(sanPham.getMaSP(), image);
+            if(images.size()!=0){
+                for (String image : images) {
+                    anhSPDAO.addAnhSP(sanPham.getMaSP(), image);
+                }
             }
+            else
+            {
+                anhSPDAO.addAnhSP(sanPham.getMaSP(), anh);
+            }
+
+
             System.out.println(sanPham.getAnh());
+
+            mess ="Tạo thành công";
 
         } else {
             SanPham sanPham = sanPhamDAO.getProductById(Integer.parseInt(maSP));
 
-            String[] arrOfStr = sanPham.getAnh().split("/");
-            String folder = sanPham.getAnh().replace("/"+arrOfStr[arrOfStr.length-1],"");
-            FileUtils.deleteDirectory(new File(request.getServletContext().getRealPath(folder)));
-
+//            if(images.size() != 0)
+//            {
+//                String[] arrOfStr = sanPham.getAnh().split("/");
+//                String folder = sanPham.getAnh().replace("/"+arrOfStr[arrOfStr.length-1],"");
+//                FileUtils.deleteDirectory(new File(request.getServletContext().getRealPath(folder)));
+//            }
 
             sanPhamDAO.updateSanPham(maDM, tensanpham, motasanpham, giagoc, giabanthuong, giakhuyenmai, soluong, anh, motangan, maSP);
             sanPham = sanPhamDAO.getProductById(Integer.parseInt(maSP));
             AnhSPDAO anhSPDAO = new AnhSPDAO();
-            anhSPDAO.deleteAnhSP(sanPham.getMaSP());
-            for (String image : images) {
-                anhSPDAO.addAnhSP(sanPham.getMaSP(), image);
+
+            if(images.size() != 0)
+            {
+                anhSPDAO.deleteAnhSP(sanPham.getMaSP());
+                for (String image : images) {
+                    anhSPDAO.addAnhSP(sanPham.getMaSP(), image);
+                    System.out.println("zo");
+                }
+
             }
+//            else {
+//                anhSPDAO.addAnhSP(sanPham.getMaSP(), anh);
+//            }
+
+            mess ="Sửa thành công";
         }
+
+        HttpSession session = request.getSession();
+        session.setAttribute("getAlert",mess);
         response.sendRedirect("Ad_ProductControl");
+        // thong bao
     }
 }
